requirements.txt
pillow
tensorflow
flask
